Move source files to src/ directory
Renamed and relocated core source files from top-level directories to the src/ directory for improved project organization. Updated build.zig to reference new file paths and fixed an include path in c_emit.c.
