+++ build.zig
@@ -25,15 +25,15 @@ pub fn build(b: *std.Build) void {
         "src/sem/symbol.c",
         "src/sem/type.c",
         "src/sem/infer.c",
-        "ir/ir.c",
-        "cfg/cfg.c",
-        "ssa/ssa.c",
-        "opt/pipeline.c",
-        "opt/sccp.c",
-        "opt/dce.c",
-        "opt/value_numbering.c",
-        "opt/licm.c",
-        "codegen/c_emit.c",
+        "src/ir/ir.c",
+        "src/cfg/cfg.c",
+        "src/ssa/ssa.c",
+        "src/opt/pipeline.c",
+        "src/opt/sccp.c",
+        "src/opt/dce.c",
+        "src/opt/value_numbering.c",
+        "src/opt/licm.c",
+        "src/codegen/c_emit.c",
         "src/codegen/codegen.c",
     }, .flags = &.{
         "-std=c11",
