+++ src/opt/sccp.c
