+++ src/opt/value_numbering.c
